make indexes and sizes uint64_t instead of int

+++ include/algo.h
@@ -5,6 +5,7 @@
 #include <stdio.h>
 #include <errno.h>
 #include <stdint.h>
+#include <inttypes.h>
 
 /*
 * NOTE about data structures:
@@ -28,11 +29,11 @@ struct dyndata_t
 
 typedef struct
 {
-    void*   structure;                      //pointer to the root of the whole structure
-    long    nbelements;                     //number of elements in the structure
-    int     elementsize;                    //size of a core element (actual data without algorithmic overlay)
-    int     (*doCompare)(void*, void*);     //comparison method
-    void    (*doPError)(char* msg, ...);    //error printing method (can be ignored if set as NULL)
+    void*       structure;                      //pointer to the root of the whole structure
+    uint64_t    nbelements;                     //number of elements in the structure
+    uint64_t    elementsize;                    //size of a core element (actual data without algorithmic overlay)
+    int         (*doCompare)(void*, void*);     //comparison method
+    void        (*doPError)(char* msg, ...);    //error printing method (can be ignored if set as NULL)
 } meta_t;
 
 // variables to use in display_AVL_tree()
@@ -48,8 +49,8 @@ typedef enum {FIRST, ANY} e_search;
 dyndata_t* allocate_dyn(meta_t* meta, void* elem);
 int free_dyn(dyndata_t* elem);
 int swap_dyn(dyndata_t* a, dyndata_t* b);
-void* get_arrayelem(meta_t* meta, int i);
-int set_arrayelem(meta_t* meta, int i, void* elem);
+void* get_arrayelem(meta_t* meta, uint64_t i);
+int set_arrayelem(meta_t* meta, uint64_t i, void* elem);
 
 //Array, list and AVL transformation
 int listToArray(meta_t* dList, meta_t* dArray, e_listtoarray action);
@@ -57,8 +58,8 @@ int arrayToList(meta_t* dArray, meta_t* dList, e_listtoarray action);
 int arrayToAVL(meta_t* dArray, meta_t* dAVL, e_listtoarray action);
 
 //Sorting algorithms
-int bubbleSortArray(meta_t*, int);
-int bubbleSortList(meta_t*, int);
+int bubbleSortArray(meta_t*, uint64_t);
+int bubbleSortList(meta_t*, uint64_t);
 int quickSortPartitioning(meta_t*, long, long);
 int quickSortArray(meta_t*, long, long);
 
@@ -66,7 +67,7 @@ int quickSortArray(meta_t*, long, long);
 int binarySearchArray(meta_t*, void*, e_search scope);
 
 //Dynamic lists
-void* get_listelem(meta_t* meta, int i);
+void* get_listelem(meta_t* meta, uint64_t i);
 dyndata_t* getright(dyndata_t* cur);
 void* getdata(dyndata_t* cur);
 int insertListTop(meta_t*, void*);
